+++ RadixWallet/Clients/OverlayWindowClient/OverlayWindowClient+Interface.swift
@@ -64,7 +64,7 @@ extension OverlayWindowClient {
 	public typealias ScheduleAlert = @Sendable (Item.AlertState) async -> Item.AlertAction
 	public typealias ScheduleAlertAndIgnoreAction = @Sendable (Item.AlertState) -> Void
 	public typealias ScheduleHUD = @Sendable (Item.HUD) -> Void
-	public typealias ScheduleSheet = @Sendable (SheetOverlayCoordinator.Root.State, SheetBehavior) -> Void
+	public typealias ScheduleSheet = @Sendable (SheetOverlayCoordinator.Root.State) -> Void
 	public typealias ScheduleFullScreen = @Sendable (FullScreenOverlayCoordinator.State) async -> FullScreenAction
 	public typealias SendAlertAction = @Sendable (Item.AlertAction, Item.AlertState.ID) -> Void
 	public typealias SendFullScreenAction = @Sendable (FullScreenAction, FullScreenID) -> Void
@@ -122,13 +122,8 @@ extension OverlayWindowClient {
 
 		case hud(HUD)
 		case alert(AlertState)
-		case sheet(SheetOverlayCoordinator.Root.State, SheetBehavior)
+		case sheet(SheetOverlayCoordinator.Root.State)
 		case fullScreen(FullScreenOverlayCoordinator.State)
-	}
-
-	public enum SheetBehavior: Sendable {
-		case enqueue
-		case replace
 	}
 }
