+++ RadixWallet/Features/AppFeature/Overlay/FullScreenOverlayCoordinator+Reducer.swift
@@ -1,7 +1,3 @@
-import ComposableArchitecture
-
-// import SwiftUI
-
 public struct FullScreenOverlayCoordinator: Sendable, FeatureReducer {
 	public struct State: Sendable, Hashable, Identifiable {
 		public let id: UUID = .init()
